refactor(block): moving prototype method to struct declaration file
