--- modules/assets/internal/block/prototype_test.go
@@ -1,27 +1,0 @@
-// Copyright [2021] - [2022], AssetMantle Pte. Ltd. and the code contributors
-// SPDX-License-Identifier: Apache-2.0
-
-package block
-
-import (
-	"github.com/AssetMantle/modules/schema/helpers"
-	"reflect"
-	"testing"
-)
-
-func TestPrototype(t *testing.T) {
-	tests := []struct {
-		name string
-		want helpers.Block
-	}{
-
-		{"+ve", block{}},
-	}
-	for _, tt := range tests {
-		t.Run(tt.name, func(t *testing.T) {
-			if got := Prototype(); !reflect.DeepEqual(got, tt.want) {
-				t.Errorf("Prototype() = %v, want %v", got, tt.want)
-			}
-		})
-	}
-}
