--- modules/splits/internal/block/prototype.go
@@ -1,10 +1,0 @@
-// Copyright [2021] - [2022], AssetMantle Pte. Ltd. and the code contributors
-// SPDX-License-Identifier: Apache-2.0
-
-package block
-
-import "github.com/AssetMantle/modules/schema/helpers"
-
-func Prototype() helpers.Block {
-	return block{}
-}
